Backend Szkielet v8 guestController

+++ Backend/src/main/java/com/PZ/AnkietBud/repositories/GuestRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Repository
 public interface GuestRepository extends CrudRepository<Guest, Long> {
     List<Guest> findAll();
-    Guest findById(long id);
+    Guest findById(Integer id);
 }
